services/audio/requirements.txt
Flask==3.0.0
requests==2.31.0
pyalsaaudio==0.10.0